build & review fixes. Movded Default into same location. Moved Mempool into config module. updated usages in jormungandr
