+++ jormungandr-lib/src/interfaces/config/mempool.rs
@@ -13,3 +13,13 @@ pub struct Mempool {
     /// interval between 2 garbage collection check of the mempool and the log cache.
     pub garbage_collection_interval: Duration,
 }
+
+impl Default for Mempool {
+    fn default() -> Self {
+        Mempool {
+            fragment_ttl: Duration::new(30 * 60, 0),
+            log_ttl: Duration::new(3600, 0),
+            garbage_collection_interval: Duration::new(3600 / 4, 0),
+        }
+    }
+}
