+++ jormungandr-lib/src/interfaces/config/mod.rs
@@ -1,0 +1,3 @@
+mod mempool;
+
+pub use mempool::Mempool;
