+++ jormungandr-lib/src/interfaces/mod.rs
@@ -3,10 +3,10 @@ mod address;
 mod block0_configuration;
 mod blockdate;
 mod certificate;
+mod config;
 mod fragment_log;
 mod leadership_log;
 mod linear_fee;
-mod mempool;
 mod old_address;
 mod settings;
 mod transaction_input;
@@ -23,10 +23,10 @@ pub use self::certificate::{
     Certificate, CertificateFromBech32Error, CertificateFromStrError, CertificateToBech32Error,
     SignedCertificate, CERTIFICATE_HRP, SIGNED_CERTIFICATE_HRP,
 };
+pub use self::config::*;
 pub use self::fragment_log::{FragmentLog, FragmentOrigin, FragmentStatus};
 pub use self::leadership_log::{EnclaveLeaderId, LeadershipLog, LeadershipLogId};
 pub use self::linear_fee::LinearFeeDef;
-pub use self::mempool::Mempool;
 pub use self::old_address::OldAddress;
 pub use self::settings::*;
 pub use self::transaction_input::{TransactionInput, TransactionInputType};
